# Faster iszero for symmetric tensors